+++ lib/units/test/unit_test.cpp
@@ -31,6 +31,15 @@ TEST(UnitTest, construct_bad_string) {
   EXPECT_THROW(Unit("abcde"), except::UnitError);
 }
 
+TEST(UnitTest, custom_unit_strings_get_rejected) {
+  // Custom (counting) units and equation units are rejected.
+  for (const auto &str : {"CXUN[0]", "CXUN[51]", "CXUN[1023]", "CXCUN[0]",
+                          "CXCUN[15]", "decibels", "ln(x)", "EQXUN[1]",
+                          "EQXUN[23]", "{corn}", "{CXCOMM[105]}"}) {
+    EXPECT_THROW_DISCARD(Unit(str), except::UnitError);
+  }
+}
+
 TEST(UnitTest, overflows) {
   // These would run out of bits in llnl/units and wrap, ensure that scipp
   // prevents this and throws instead.
@@ -238,7 +247,7 @@ TEST(UnitParseTest, singular_plural) {
 TEST(UnitFormatTest, roundtrip_string) {
   for (const auto &s :
        {"m", "m/s", "meV", "pAh", "mAh", "ns", "counts", "counts^2",
-        "counts/meV", "1/counts", "counts/m", "Y", "M", "D"}) {
+        "counts/meV", "1/counts", "counts/m", "rad", "$", "Y", "M", "D"}) {
     const auto unit = units::Unit(s);
     EXPECT_EQ(to_string(unit), s);
     EXPECT_EQ(units::Unit(to_string(unit)), unit);